Switch to chef 12.5 custom resource
